Refactor, move test helper to its own package
