+++ algo/helper/helper.go
@@ -1,4 +1,4 @@
-package algo
+package helper
 
 import (
 	"math/rand"
